requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2017.1.23
future==0.16.0
python-telegram-bot==5.3.0
requests==2.13.0
urllib3==1.20
